Move is_sdi_addable to substanced.folder
