+++ substanced/folder/tests/test_init.py
@@ -1141,6 +1141,57 @@ class TestCopyHook(unittest.TestCase):
         inst = self._makeOne(child)
         self.assertRaises(ResumeCopy, inst, parent, None)
 
+
+class Test_is_sdi_addable(unittest.TestCase):
+    def setUp(self):
+        self.config = testing.setUp()
+
+    def tearDown(self):
+        testing.tearDown()
+
+    def call_it(self, context, request, content_type):
+        from .. import is_sdi_addable
+        return is_sdi_addable(context, request, content_type)
+
+    def test_true_if_sdi_addable_is_none(self):
+        config = self.config
+        config.include('substanced.content')
+        config.add_content_type('Some Type', object)
+        context = testing.DummyResource(__sdi_addable__=None)
+        request = testing.DummyRequest()
+        self.assertTrue(self.call_it(context, request, 'Some Type'))
+
+    def test_with_callable_sdi_addable(self):
+        config = self.config
+        config.include('substanced.content')
+        config.add_content_type('Type1', object, addable=True)
+        config.add_content_type('Type2', object, addable=False)
+
+        def addable(context, intr):
+            meta = intr['meta']
+            return meta.get('addable', False)
+
+        context = testing.DummyResource(__sdi_addable__=addable)
+        request = testing.DummyRequest()
+        self.assertTrue(self.call_it(context, request, 'Type1'))
+        self.assertFalse(self.call_it(context, request, 'Type2'))
+
+    def test_with_sequence_sdi_addable(self):
+        config = self.config
+        config.include('substanced.content')
+        config.add_content_type('Type1', object)
+        config.add_content_type('Type2', object)
+
+        context = testing.DummyResource(__sdi_addable__=('Type1',))
+        request = testing.DummyRequest()
+        self.assertTrue(self.call_it(context, request, 'Type1'))
+        self.assertFalse(self.call_it(context, request, 'Type2'))
+
+    def test_false_if_content_type_unknown(self):
+        context = testing.DummyResource()
+        request = testing.DummyRequest()
+        self.assertFalse(self.call_it(context, request, 'Unknown'))
+
 class Test_AddableContentPredicate(unittest.TestCase):
     def _makeOne(self, val, config):
         from .. import _AddableContentPredicate
