+++ substanced/folder/tests/test_util.py
@@ -1,7 +1,6 @@
 
 import unittest
 
-from pyramid import testing
 
 class Test_slugify_in_context(unittest.TestCase):
     def _callFUT(self, context, name, remove_extension=True):
@@ -63,54 +62,3 @@ class Test_slugify_in_context(unittest.TestCase):
             self._callFUT(context, 'boo.pdf', remove_extension=False),
             'boo-pdf-3'
             )
-
-
-class Test_is_sdi_addable(unittest.TestCase):
-    def setUp(self):
-        self.config = testing.setUp()
-
-    def tearDown(self):
-        testing.tearDown()
-
-    def call_it(self, context, request, content_type):
-        from ..util import is_sdi_addable
-        return is_sdi_addable(context, request, content_type)
-
-    def test_true_if_sdi_addable_is_none(self):
-        config = self.config
-        config.include('substanced.content')
-        config.add_content_type('Some Type', object)
-        context = testing.DummyResource(__sdi_addable__=None)
-        request = testing.DummyRequest()
-        self.assertTrue(self.call_it(context, request, 'Some Type'))
-
-    def test_with_callable_sdi_addable(self):
-        config = self.config
-        config.include('substanced.content')
-        config.add_content_type('Type1', object, addable=True)
-        config.add_content_type('Type2', object, addable=False)
-
-        def addable(context, intr):
-            meta = intr['meta']
-            return meta.get('addable', False)
-
-        context = testing.DummyResource(__sdi_addable__=addable)
-        request = testing.DummyRequest()
-        self.assertTrue(self.call_it(context, request, 'Type1'))
-        self.assertFalse(self.call_it(context, request, 'Type2'))
-
-    def test_with_sequence_sdi_addable(self):
-        config = self.config
-        config.include('substanced.content')
-        config.add_content_type('Type1', object)
-        config.add_content_type('Type2', object)
-
-        context = testing.DummyResource(__sdi_addable__=('Type1',))
-        request = testing.DummyRequest()
-        self.assertTrue(self.call_it(context, request, 'Type1'))
-        self.assertFalse(self.call_it(context, request, 'Type2'))
-
-    def test_false_if_content_type_unknown(self):
-        context = testing.DummyResource()
-        request = testing.DummyRequest()
-        self.assertFalse(self.call_it(context, request, 'Unknown'))
